requirements.txt
aiogram==3.15.0
pyyaml==6.0.2
python-dotenv==1.0.1
aiofiles==24.1.0